structure fix
