+++ src/main/java/ru/practicum/shareit/comment/CommentDto.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit.comment.dto;
+package ru.practicum.shareit.comment;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
