+++ src/main/java/ru/practicum/shareit/comment/CommentMapper.java
@@ -1,8 +1,6 @@
-package ru.practicum.shareit.comment.mapper;
+package ru.practicum.shareit.comment;
 
 import org.springframework.stereotype.Component;
-import ru.practicum.shareit.comment.dto.CommentDto;
-import ru.practicum.shareit.comment.model.Comment;
 
 @Component
 public class CommentMapper {
