+++ src/main/java/ru/practicum/shareit/comment/CommentRepository.java
@@ -1,8 +1,7 @@
-package ru.practicum.shareit.comment.repository;
+package ru.practicum.shareit.comment;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import ru.practicum.shareit.comment.model.Comment;
 
 import java.util.List;
