+++ src/main/java/ru/practicum/shareit/item/ItemMapper.java
@@ -1,7 +1,4 @@
-package ru.practicum.shareit.item.mapper;
-
-import ru.practicum.shareit.item.dto.ItemDto;
-import ru.practicum.shareit.item.model.Item;
+package ru.practicum.shareit.item;
 
 public class ItemMapper {
     public static ItemDto toItemDto(Item item) {
