+++ src/main/java/ru/practicum/shareit/item/ItemRepository.java
@@ -1,8 +1,7 @@
-package ru.practicum.shareit.item.repository;
+package ru.practicum.shareit.item;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import ru.practicum.shareit.item.model.Item;
 
 import java.util.Collection;
