+++ src/main/java/ru/practicum/shareit/item/service/ItemService.java
@@ -1,7 +1,7 @@
 package ru.practicum.shareit.item.service;
 
-import ru.practicum.shareit.comment.dto.CommentDto;
-import ru.practicum.shareit.item.dto.ItemDto;
+import ru.practicum.shareit.comment.CommentDto;
+import ru.practicum.shareit.item.ItemDto;
 
 import java.util.Collection;
