+++ src/main/java/ru/practicum/shareit/request/ItemRequestMapper.java
@@ -1,7 +1,4 @@
-package ru.practicum.shareit.request.mapper;
-
-import ru.practicum.shareit.request.dto.ItemRequestDto;
-import ru.practicum.shareit.request.model.ItemRequest;
+package ru.practicum.shareit.request;
 
 public class ItemRequestMapper {
     public static ItemRequestDto toItemRequestDto(ItemRequest itemRequest) {
