+++ src/main/java/ru/practicum/shareit/user/UserRepository.java
@@ -1,8 +1,7 @@
-package ru.practicum.shareit.user.repository;
+package ru.practicum.shareit.user;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import ru.practicum.shareit.user.model.User;
 
 import java.util.Optional;
